+++ src/commands/admin/index.ts
@@ -2,7 +2,6 @@ import { configuration } from "./configuration";
 import { exportData } from "./export";
 import { bulkAdd, bulkAddTemplate } from "./import";
 import { generateTemplate, registerTemplate } from "./template";
-
 export const ADMIN = [
 	configuration,
 	generateTemplate,