+++ src/commands/gimmick/index.ts
@@ -1,4 +1,5 @@
 import { displayUser } from "./display";
 import { graph } from "./graph";
+import { editAvatar } from "./edit_avatar";
 
-export const GIMMICK = [displayUser, graph];
+export const GIMMICK = [displayUser, graph, editAvatar];